+++ tests/testthat/test-NanoStringQC.R
@@ -24,3 +24,17 @@ test_that("Error if positive controls don't have concentrations in brackets", {
     gsub("\\([0-9].*", "\\1", hlo.r$Name[hlo.r$Code.Class == "Positive"])
   expect_error(NanoStringQC(hlo.r, expQC[expQC$geneRLF == "HL1", ]))
 })
+
+test_that("Error if non-standard column names", {
+  hlo.r.2 <- hlo.r
+  names(hlo.r.2)[1:2] <- c("A", "B")
+  expect_error(NanoStringQC(hlo.r.2, expQC[expQC$geneRLF == "HL1", ]))
+})
+
+test_that("Error if missing endogenous or housekeeping genes", {
+  hlo.r.3 <- hlo.r[hlo.r$Code.Class %in% c("Endogenous", "Negative", "Positive"), ]
+  expect_error(NanoStringQC(hlo.r.3, expQC[expQC$geneRLF == "HL1", ]))
+
+  hlo.r.4 <- hlo.r[hlo.r$Code.Class %in% c("Housekeeping", "Negative", "Positive"), ]
+  expect_error(NanoStringQC(hlo.r.4, expQC[expQC$geneRLF == "HL1", ]))
+})