+++ NEWS.md
@@ -1,3 +1,5 @@
+# nanostringr (development version)
+
 # nanostringr 0.1.2
 
 * sort sample names in HLD and OVD cohorts by numeric order, not lexicographic order