+++ NEWS.md
@@ -1,6 +1,6 @@
 # nanostringr (development version)
 
-* unit tests for `read_rcc()`
+* unit tests for `read_rcc()` and internal functions `check_colnames()`, `check_genes()`
 
 * fix bug using `NanoStringQC()` for single sample data